Fix Cb 16 (Rotate left through carry)

+++ src/RetroEmu.Devices/DMG/CPU/Processor.Operations.cs
@@ -422,7 +422,7 @@ public partial class Processor
         var newCarry = (input & 0x80) > 0;
         var lsbMask = IsSet(Flag.Carry) ? 0x01 : 0x00;
 
-        var result = (input << 1) | lsbMask;
+        var result = (byte)((input << 1) | lsbMask);
 
         SetFlagToValue(Flag.Zero, result == 0);
         ClearFlag(Flag.Subtract);
